Add files via upload
The project's aim is to provide help in situation faced regularly by accountants:

I have a list of values (let's say 20 values) and an another value which is probably a sum of 2
 or more values out of these 20 values. I have no idea how many values are aggregated. 
I want to get the list of values that got added.

The tool takes a series of values and the aggregated value as input and searches for the "ingredients" 
of the sum. The result is the list of values that got added. It may find more than one matching list.

The project is rather a beta version - the GUI needs to be improved. Performance wise it can not work
 on very long list of values as the amount of calculation inside is enormous.